functions/requirements.txt
azure-functions
azure-servicebus>=7.11.0
